Added PieChart and input field history control

+++ src/main/java/utils/TasksCounter.java
@@ -20,15 +20,15 @@ public class TasksCounter {
     /**
      * Gets percentage of tasks marked "done" in given set of tasks
      *
-     * @return integer of percentage marked done
+     * @return float value of percentage marked done
      */
-    public int getPercCompleted() {
+    public float getPercCompleted() {
         float totalCompleted = 0;
         for (int i = 0; i < tasks.size(); i++) {
             if (tasks.get(i).getIsDone()) {
                 totalCompleted += 1;
             }
         }
-        return (int) (totalCompleted / (float) tasks.size() * 100f);
+        return totalCompleted / (float) tasks.size() * 100f;
     }
 }
